% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{collect_input_shiny}
\alias{collect_input_shiny}
\title{Append sequence input to global input dataframe}
\usage{
collect_input_shiny(input, global_value)
}
\arguments{
\item{input}{}

\item{global_value}{}
}
\value{

}
\description{
Internal use in the Shiny UI only.
}
