% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_design.R
\name{create_primer_pair}
\alias{create_primer_pair}
\title{Design primer pair from overlaps.}
\usage{
create_primer_pair(overlaps)
}
\arguments{
\item{overlaps}{A dataframe containing the overlap information for both input
types. The \code{overlaps} object is created by \code{get_overlaps}.}
}
\value{
A dataframe containing the forward and reverse primer sequences.
}
\description{
This function uses the overlap information in the \code{overlaps} object
to design a primer pair by concatenating the overlap sequences in the
correct order (see Details).
}
\details{
In order to design the primers for the first PCR step of LambdaPCR, where we
add template-overlaping flanking regions to our target gene sequence, 
we have to use the following rules:

\itemize{
  \item \strong{Forward primer:} left overlap + target left overlap, 
  5'-3' orientation
  \item \strong{Reverse primer:} complement of target right overlap + 
  complement of template right overlap,  5'-3' orientation
}
}
