% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_design.R
\name{create_primers}
\alias{create_primers}
\title{Create primers from target and vector DNA sequences.}
\usage{
create_primers(run_object)
}
\arguments{
\item{run_object}{}
}
\description{
Create primers from target and vector DNA sequences.
}
