% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_design.R
\name{get_annealing_regions}
\alias{get_annealing_regions}
\title{Extract target or vector-annealing regions from Template sequences.}
\usage{
get_annealing_regions(template, position = NULL, min_length = NULL,
  max_length = NULL, ...)
}
\arguments{
\item{template}{}

\item{position}{}

\item{min_length}{}

\item{max_length}{}

\item{...}{}
}
\description{
Extract target or vector-annealing regions from Template sequences.
}
