% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_imports.R
\docType{package}
\name{lambdaPrimeR}
\alias{lambdaPrimeR}
\alias{lambdaPrimeR-package}
\title{lambdaPrimeR: A package for designing primers for Lambda PCR}
\description{
lambdaPrimeR: A package for designing primers for Lambda PCR
}
