% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamic_calculations.R
\name{nupack_args}
\alias{nupack_args}
\title{Gather NUPACK's tools parameters.}
\usage{
nupack_args(tool, input_sequence, melting_temperature, multi,
  dangles = c("some", "none", "all"))
}
\arguments{
\item{tool}{Name of the NUPACK tool.}

\item{input_sequence}{A string corresponding to the input sequence.}

\item{melting_temperature}{A numeric value indicating the melting temperature
to be used during the NUPACK's calculations.}

\item{multi}{Multiple input sequences? (Default is \code{FALSE}).}

\item{dangles}{A string with the way in which dangle energies are 
incorporated. Possible values: 

\itemize{
  \item \code{none}: No dangle energy are incorporated
  \item \code{some} (default): A dangle energy is incorporated for each 
  unpaired base flanking a duplex (a base flanking two duplexes contributes 
  only the minimum of the two possible dangle energies).
  \item \code{all}: A dangle energy is incorporated for each base flanking a 
  duplex regardless of whether it is paired.
}}
}
\value{
A data frame with base parameters for a NUPACK function
}
\description{
Gather NUPACK's tools parameters.
}
