% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamic_calculations.R
\name{nupack_mfe}
\alias{nupack_mfe}
\title{Minimum free energy (MFE) structure determination using NUPACK's \code{mfe} 
tool.}
\usage{
nupack_mfe(input_sequence, melting_temperature, multi = FALSE,
  dangles = c("some", "none", "all"), degenerate = FALSE)
}
\arguments{
\item{input_sequence}{A string corresponding to the input sequence.}

\item{melting_temperature}{A numeric value indicating the melting temperature
to be used during the NUPACK's calculations.}

\item{multi}{Multiple input sequences? (Default is \code{FALSE}).}

\item{dangles}{A string with the way in which dangle energies are 
incorporated. Possible values: 

\itemize{
  \item \code{none}: No dangle energy are incorporated
  \item \code{some} (default): A dangle energy is incorporated for each 
  unpaired base flanking a duplex (a base flanking two duplexes contributes 
  only the minimum of the two possible dangle energies).
  \item \code{all}: A dangle energy is incorporated for each base flanking a 
  duplex regardless of whether it is paired.
}}

\item{degenerate}{Compute all structures with MFE? Default is \code{FALSE}.}
}
\value{
A data frame with the MFE value and the dot-parens-plus drawing of 
MFE structure.
}
\description{
Minimum free energy (MFE) structure determination using NUPACK's \code{mfe} 
tool.
}
