% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamic_calculations.R
\name{nupack_pfunc}
\alias{nupack_pfunc}
\title{Calculate the partition function of a DNA complex using NUPACK's \code{pfunc}
tool.}
\usage{
nupack_pfunc(input_sequence, melting_temperature, multi = FALSE,
  dangles = c("some", "none", "all"))
}
\arguments{
\item{input_sequence}{A string corresponding to the input sequence.}

\item{melting_temperature}{A numeric value indicating the melting temperature
to be used during the NUPACK's calculations.}

\item{multi}{Multiple input sequences? (Default is \code{FALSE}).}

\item{dangles}{A string with the way in which dangle energies are 
incorporated. Possible values: 

\itemize{
  \item \code{none}: No dangle energy are incorporated
  \item \code{some} (default): A dangle energy is incorporated for each 
  unpaired base flanking a duplex (a base flanking two duplexes contributes 
  only the minimum of the two possible dangle energies).
  \item \code{all}: A dangle energy is incorporated for each base flanking a 
  duplex regardless of whether it is paired.
}}
}
\value{
A data frame containing the free energy of the complex and the
partition function.
}
\description{
Calculate the partition function of a DNA complex using NUPACK's \code{pfunc}
tool.
}
