% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primers.R
\docType{class}
\name{Primers-class}
\alias{Primers-class}
\title{S4 class representing a group of lambda and theta primers.}
\value{
The \code{Primers} constructor return an object of class \code{Primers},
an instance of data frame.
}
\description{
The \code{Primers} class contains the information about the primers designed 
to amplify and insert the target sequence into the vector (the lambda primers),
and the small primer that will be used to amplify the other vector strand
(the theta primer).
}
