% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_ui.R
\name{run_app_ui}
\alias{run_app_ui}
\title{Open lambdaPrimeR Shiny user-interface}
\usage{
run_app_ui()
}
\value{
Opens Shiny app on a new browser window
}
\description{
Open lambdaPrimeR Shiny user-interface
}
