% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamic_calculations.R
\name{run_nupack}
\alias{run_nupack}
\title{Execute a NUPACK command-line tool and retrieve outputs.}
\usage{
run_nupack(input_file, melting_temperature, exec, timed = FALSE,
  multi = FALSE, paired = TRUE)
}
\arguments{
\item{input_file}{The path or a vector containing the path to the NUPACK 
input file (suffix ".in").}

\item{melting_temperature}{A numeric value corresponding to the melting 
temperature to be used during NUPACK's calculations.}

\item{exec}{A string indicating the name of the NUPACK command to be run: 
"\code{pfunc}", "\code{pairs}" or "\code{mfe}".}

\item{timed}{Time command runtime? Default is \code{FALSE}.}

\item{multi}{Multiple input sequences? Default is \code{FALSE}.}

\item{paired}{Return base pair formation probabilities from \code{pairs}
command? Default it \code{TRUE}.}
}
\value{
A data frame containing different fields depending on the NUPACK
command chosen with the \code{exec} parameter.
}
\description{
Execute a NUPACK command-line tool and retrieve outputs.
}
