% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Template_classes.R
\docType{class}
\name{Template-class}
\alias{Template-class}
\title{S4 class representing a DNA sequence template.}
\value{
The \code{Template} constructor returns an object of class \code{Template},
an instance of a data frame.
}
\description{
This is the master class for the \code{lambdaPrimeR} package. It represents
a input DNA sequence that can either be the Target or the Vector sequence.
Both the target and vector sequences will be used as templates for the design 
of the lambda PCR primers.
}
\details{
This master-class is internal, as the two subclasses \code{Target} and
\code{Vector} will be the ones used to design the primers in this package.
}
