% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Template_classes.R
\name{validate_template}
\alias{validate_template}
\title{Validade a \code{Template} object.}
\usage{
validate_template(object)
}
\arguments{
\item{object}{The input data frame (returned by \code{read_sequences}) to be 
checked against a template data frame.}
}
\value{
\code{TRUE} if object is a valid data frame, otherwise return a list
of detected errors during input validation.
}
\description{
Checks if a \code{Template} object or an object of any subclass is valid 
or not.
}
\keyword{interal}
