% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer_design.R
\name{melt_temperature}
\alias{melt_temperature}
\title{Calculate primer melting temperature.}
\usage{
melt_temperature(sequence, template_conc, ion_conc)
}
\arguments{
\item{sequence}{A string corresponding to the primer sequence.}

\item{template_conc}{A numeric value corresponding to the DNA template 
concentration. Defauts to $10^-4$ M.}

\item{ion_conc}{A string corresponding to the concentration of ion. See 
MELTING documentation for details.}
}
\value{
A dataframe containing the primer sequence and the calculated melting
temperature \code{tm}.
}
\description{
Calculate primer melting temperature.
}
