% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Template_classes.R
\name{read_sequences}
\alias{read_sequences}
\title{Read sequences from file.}
\usage{
read_sequences(file, input_type = c("target", "vector"),
  header_structure = NULL, sep = NULL, name_field = NULL)
}
\arguments{
\item{file}{A path to a FASTA file, a vector containing the file path or a
string containing the sequence with header.}

\item{input_type}{A string used to indicate if the sequence is a 'target' or a
'template'.}

\item{header_structure}{A character vector corresponding to the structure of 
sequence header line. Each vector element should correspond to a field in the 
header line, delimitated by \code{sep}.}

\item{sep}{A single character or regex used to separate fields in the sequence
header according to the elements of \code{header_structure}. Defaults to '\\|'.}

\item{name_field}{A string inticating what field of the header string should 
be used as the sequence \code{name}.}
}
\value{
A dataframe.
}
\description{
\code{read_sequences} reads a FASTA file into a dataframe, one sequence per
row. This function also adds a \code{type} tag to each sequence, indicating if
they are either a \code{target} gene or a \code{vector} sequence. Depending on
the value of \code{type}, an instance of either the \code{Target} or 
\code{Vector} class will be created. User can optionally select what part of 
the sequence header will be used as the sequence \code{name}.
}
\examples{
header <- c("name", "id", "extra_info")
read_sequences("data/test.fasta", input_type = 'target')
read_sequences("data/test.fasta", input_type = 'vector', header_structure = header, separator = '-')
}
